+++ package.json
@@ -41,12 +41,12 @@
   "dependencies": {
     "@loadable/component": "^5.10.2",
     "@mdx-js/react": "^1.5.2",
-    "@reach/alert": "^0.1.5",
-    "@reach/combobox": "^0.8.5",
-    "@reach/dialog": "^0.4.0",
-    "@reach/menu-button": "^0.10.0",
-    "@reach/popover": "^0.9.0",
-    "@reach/tooltip": "^0.8.5",
+    "@reach/alert": "^0.10.1",
+    "@reach/combobox": "^0.10.1",
+    "@reach/dialog": "^0.10.1",
+    "@reach/menu-button": "^0.10.1",
+    "@reach/popover": "^0.10.1",
+    "@reach/tooltip": "^0.10.1",
     "@types/lodash.sample": "^4.2.6",
     "case": "^1.6.2",
     "date-fns": "^2.8.1",
