--- src/components/Dropdown/index.ts
@@ -1,1 +1,0 @@
-export * from "./Dropdown"
