+++ src/components/DropdownMenu/DropdownMenu.tsx
@@ -1,0 +1,81 @@
+/** @jsx jsx */
+import { jsx } from "@emotion/core"
+import React from "react"
+import {
+  Menu,
+  MenuProps,
+  MenuPopover,
+  MenuPopoverProps,
+  MenuList,
+  MenuListProps,
+  MenuButton,
+  MenuButtonProps,
+  MenuItem,
+  MenuItemProps,
+  MenuLink,
+  MenuLinkProps,
+  useMenuButtonContext,
+} from "@reach/menu-button"
+import { MdKeyboardArrowDown } from "react-icons/md"
+import {
+  dropdownCss,
+  dropdownButtonCss,
+  menuItemCss,
+  dropdownButtonIconCss,
+} from "./DropdownMenu.styles"
+import { DisableReachStyleCheck } from "../../utils/helpers/DisableReachStyleCheck"
+
+export type DropdownMenuProps = MenuProps
+
+export const DropdownMenu: React.FC<DropdownMenuProps> = props => (
+  <React.Fragment>
+    <DisableReachStyleCheck reachComponent="menu-button" />
+    <Menu {...props} />
+  </React.Fragment>
+)
+
+export type DropdownMenuButtonProps = MenuButtonProps
+
+export const DropdownMenuButton: React.FC<DropdownMenuButtonProps> = props => (
+  <MenuButton {...props} />
+)
+
+export type DropdownMenuButtonStyledProps = DropdownMenuButtonProps
+
+export const DropdownMenuButtonStyled: React.FC<DropdownMenuButtonProps> = ({
+  children,
+  ...props
+}) => (
+  <DropdownMenuButton {...props} css={dropdownButtonCss}>
+    {children}
+    <MdKeyboardArrowDown css={dropdownButtonIconCss} />
+  </DropdownMenuButton>
+)
+
+export type DropdownMenuItemsProps = MenuListProps
+
+export const DropdownMenuItems: React.FC<DropdownMenuItemsProps> = props => (
+  <MenuList {...props} css={dropdownCss}></MenuList>
+)
+
+export type DropdownMenuItemProps = MenuItemProps
+
+export const DropdownMenuItem: React.FC<DropdownMenuItemProps> = props => (
+  <MenuItem {...props} css={menuItemCss} />
+)
+
+export type DropdownMenuLinkProps = MenuLinkProps
+
+export function DropdownMenuLink(props: DropdownMenuLinkProps) {
+  return <MenuLink {...props} />
+}
+
+export type DropdownMenuPopoverProps = MenuPopoverProps
+
+export function DropdownMenuPopover(props: DropdownMenuPopoverProps) {
+  return <MenuPopover {...props} />
+}
+
+export function useDropdownMenuContext() {
+  return useMenuButtonContext()
+}
